requirements.txt
pandas
numpy
python-dotenv
requests
fastapi
uvicorn
pydantic
streamlit
openai
pytest
openpyxl
